fix: move attestation response dto
