+++ platform-backend/backend-service/src/main/java/cn/flying/service/remote/FileRemoteClient.java
@@ -13,7 +13,7 @@ import cn.flying.platformapi.request.GetShareInfoRequest;
 import cn.flying.platformapi.request.GetUserShareCodesRequest;
 import cn.flying.platformapi.request.ShareFilesRequest;
 import cn.flying.platformapi.request.StoreAttestationBatchRequest;
-import cn.flying.platformapi.request.StoreAttestationBatchResponse;
+import cn.flying.platformapi.response.StoreAttestationBatchResponse;
 import cn.flying.platformapi.request.StoreFileRequest;
 import cn.flying.platformapi.request.StoreFileResponse;
 import cn.flying.platformapi.response.FileDetailVO;
