+++ platform-backend/backend-service/src/test/java/cn/flying/service/attestation/AttestationBatchServiceImplTest.java
@@ -13,7 +13,7 @@ import cn.flying.dao.mapper.AttestationLeafMapper;
 import cn.flying.dao.mapper.FileMapper;
 import cn.flying.platformapi.constant.Result;
 import cn.flying.platformapi.request.StoreAttestationBatchRequest;
-import cn.flying.platformapi.request.StoreAttestationBatchResponse;
+import cn.flying.platformapi.response.StoreAttestationBatchResponse;
 import cn.flying.service.remote.FileRemoteClient;
 import cn.flying.test.builders.BuilderResetExtension;
 import cn.flying.test.builders.FileTestBuilder;
